Tweaks from review.
